+++ routes/study/search.js
